modules/plugin/test/mochitest/test_pluginstream_seek_close.html

<body>
<head>
  <title>NPAPI Seekable NPStream Test</title>
  <script type="text/javascript" 
          src="/MochiKit/packed.js"></script>
  <script type="text/javascript" 
          src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" 
        href="/tests/SimpleTest/test.css" />
</head>
<body>
  <p id="display"></p>

  <iframe id="testframe" name="testframe" onload="frameLoaded()"></iframe>

  







  <embed src="neverending.sjs" streammode="seek" closestream
         frame="testframe" streamchunksize="1024" range="100,100"
         id="embedtest" style="width: 400px; height: 100px;"
         type="application/x-test"></embed>

  <script class="testbody" type="application/javascript">
  SimpleTest.waitForExplicitFinish();

  var testframe = document.getElementById('testframe');

  function frameLoaded() {
    var content = testframe.contentDocument.body.innerHTML;
    if (!content.length)
      return;

    ok(true, "We didn't crash");
    SimpleTest.finish();
  }
  </script>
</body>
</html>
